The position now used double

+++ M2107-P2.1-PiesQuest/src/Position.java
@@ -6,12 +6,12 @@ public class Position {
 	/**
 	 * The coordinate in the X axis
 	 */
-	public int x;
+	public double x;
 	
 	/**
 	 * The coordinate in the Y axis
 	 */
-	public int y;
+	public double y;
 	
 	/**
 	 * Set the position to the origin (0, 0)
@@ -25,7 +25,7 @@ public class Position {
 	 * @param theX the position in the x axis
 	 * @param theY the position in the y axis
 	 */
-	public Position(int theX, int theY) {
+	public Position(double theX, double theY) {
 		this.x = theX;
 		this.y = theY;
 	}
